refactor: move hasDnsResultOrderOverride and isWSL2 to internal lib/net-utils.mjs
Agent-Logs-Url: https://github.com/nujovich/mint/sessions/42602eac-c540-4ed8-b7a6-bb49822c56f3
