+++ lib/__tests__/prompts.test.mjs
@@ -1,5 +1,4 @@
-import { describe, it, expect, vi, afterEach, beforeEach } from 'vitest'
-import { readFileSync } from 'node:fs'
+import { describe, it, expect, vi, afterEach } from 'vitest'
 import {
   stripFences,
   resolveTarget,
@@ -8,13 +7,7 @@ import {
   buildResolvePrompt,
   buildExportPrompt,
   callAnthropic,
-  hasDnsResultOrderOverride,
-  isWSL2,
 } from '../prompts.mjs'
-
-vi.mock('node:fs', () => ({
-  readFileSync: vi.fn(),
-}))
 
 describe('stripFences', () => {
   it('strips a js-fenced code block', () => {
@@ -230,78 +223,5 @@ describe('callAnthropic', () => {
     }))
     const result = await callAnthropic({ apiKey: 'test-key', prompt: 'hello' })
     expect(result).toBe('ok')
-  })
-})
-
-describe('hasDnsResultOrderOverride', () => {
-  let originalExecArgv
-
-  beforeEach(() => {
-    originalExecArgv = process.execArgv
-    process.execArgv = []
-    vi.stubEnv('NODE_OPTIONS', '')
-  })
-
-  afterEach(() => {
-    process.execArgv = originalExecArgv
-    vi.unstubAllEnvs()
-  })
-
-  it('returns true when NODE_OPTIONS includes --dns-result-order', () => {
-    vi.stubEnv('NODE_OPTIONS', '--dns-result-order=verbatim')
-    expect(hasDnsResultOrderOverride()).toBe(true)
-  })
-
-  it('returns true when execArgv contains --dns-result-order', () => {
-    process.execArgv = ['--dns-result-order']
-    expect(hasDnsResultOrderOverride()).toBe(true)
-  })
-
-  it('returns true when execArgv contains --dns-result-order=value', () => {
-    process.execArgv = ['--dns-result-order=verbatim']
-    expect(hasDnsResultOrderOverride()).toBe(true)
-  })
-
-  it('returns false when neither NODE_OPTIONS nor execArgv signal an override', () => {
-    expect(hasDnsResultOrderOverride()).toBe(false)
-  })
-})
-
-describe('isWSL2', () => {
-  afterEach(() => {
-    vi.unstubAllEnvs()
-    vi.mocked(readFileSync).mockReset()
-  })
-
-  it('returns true when WSL_DISTRO_NAME is set', () => {
-    vi.stubEnv('WSL_DISTRO_NAME', 'Ubuntu')
-    expect(isWSL2()).toBe(true)
-  })
-
-  it('returns true when WSL_INTEROP is set', () => {
-    vi.stubEnv('WSL_DISTRO_NAME', '')
-    vi.stubEnv('WSL_INTEROP', '/run/interop')
-    expect(isWSL2()).toBe(true)
-  })
-
-  it('returns true when /proc/version contains microsoft', () => {
-    vi.stubEnv('WSL_DISTRO_NAME', '')
-    vi.stubEnv('WSL_INTEROP', '')
-    vi.mocked(readFileSync).mockReturnValue('Linux microsoft version')
-    expect(isWSL2()).toBe(true)
-  })
-
-  it('returns false when /proc/version has no microsoft', () => {
-    vi.stubEnv('WSL_DISTRO_NAME', '')
-    vi.stubEnv('WSL_INTEROP', '')
-    vi.mocked(readFileSync).mockReturnValue('Linux 5.15.0 Ubuntu')
-    expect(isWSL2()).toBe(false)
-  })
-
-  it('returns false when readFileSync throws', () => {
-    vi.stubEnv('WSL_DISTRO_NAME', '')
-    vi.stubEnv('WSL_INTEROP', '')
-    vi.mocked(readFileSync).mockImplementation(() => { throw new Error('ENOENT') })
-    expect(isWSL2()).toBe(false)
   })
 })
